Statically Cache Blog Page
This forces the Blog Page to be built and cached statically. In order to
make sure that search still works, I factored out a smaller BlogSection
component that does client side search against that data.

For some reasons, requests for the new page are still being sent to the
server, but at least the page is cached on the server now preventing any
serverless function calls to fetch the data.