[FILECOIN][UXIT-3007] Add side-by-side layout (#1560)
* Remove unused `HardDrivesIcon` import

* refactor: Introduce SideBySideContent component

This commit introduces a new reusable `SideBySideContent` component to standardize the layout of sections that feature a title, description, and call-to-action buttons.

The Homepage, Learn, and Offer Storage pages have been refactored to use this new component. This change simplifies the page structure, reduces code duplication, and ensures a consistent user interface across the site.

* refactor: Use SideBySideContent for consistent section layout

This commit refactors the Homepage, Learn, and Offer Storage pages to use a new `SideBySideContent` component.

Previously, these pages contained duplicated markup for sections with a title, description, and call-to-action buttons.

By abstracting this pattern into a reusable `SideBySideContent` component, we reduce code duplication, improve maintainability, and ensure a consistent visual layout across the site.

* Correct the import path for the `SideBySideContent` component on the Learn page.

* Update apps/filecoin-site/src/app/_styles/globals.css

Co-authored-by: Mirha Masala <26620750+mirhamasala@users.noreply.github.com>

* Remove the redundant `div` wrapper around the `Heading` component.

* refactor: Rename SideBySideContent to SplitSectionContent

* Refactor: Rename CSS classes for SplitSectionContent

Update the CSS class names for the heading and description elements from `side-by-side-*` to `split-section-*`.

---------

Co-authored-by: Mirha Masala <26620750+mirhamasala@users.noreply.github.com>