[UXIT-3141] Add Image Fallback (#1679)
* Add fallback image to graphicsData

* feat(blog): Implement fallback image for posts

Ensure all blog posts display a header image, even if one is not specified in the frontmatter.

A global fallback image is now used in the `BlogCard`, `BlogPostHeader`, and the featured post section on the blog index page.

* refactor(blog): Remove redundant image fields from frontmatter

* Refactor: Simplify blog post header image handling

Remove the hardcoded fallback image from the blog post page and simplify the `BlogPostHeader` component's props.

The responsibility for displaying a fallback image is now fully contained within the `BlogPostHeader` component. This component now accepts an optional image URL string directly, making the parent page component cleaner and centralizing the image display logic.