[FILECOIN][UXIT-3192] Fix Gradient Position (#1723)
* refactor(build): Rework header and gradient layout

Combine the page header and its background gradient into a single container on the "Build on Filecoin" page.

* refactor(ui): Create GradientContainer for hero sections

Introduces a new `GradientContainer` component to encapsulate the logic for sections that feature a large, decorative background gradient.

* refactor(case-studies): Use GradientContainer for page header

Refactor the Case Studies page header to use the new `GradientContainer` component. This replaces a custom `div` with `clsx` and standardizes the implementation of the gradient background.

* feat(components): Enhance GradientContainer for overflow clipping

The `GradientContainer` component is updated to support vertical overflow clipping with a fade-out effect.

A new `overflowY` prop has been added, which can be set to 'clip'. When clipped, a gradient is automatically applied to the bottom of the container to create a smooth visual transition for overflowing content.

The Community Hub page hero has been refactored to use this enhanced component, simplifying its structure and encapsulating the container logic. The hero background graphic has also been repositioned to align with the new layout.

* refactor: Improve GradientContainer API and reusability

The `GradientContainer` component has been updated to use a more explicit API. The `overflowY` prop is replaced with two boolean props, `clipY` and `overlay`, for clearer control over vertical clipping and the gradient overlay.

This change improves component reusability, allowing it to replace a custom implementation on the `provide-storage` page. The `community-hub` page has been updated to use the new props.

---------

Co-authored-by: Mirha Masala <26620750+mirhamasala@users.noreply.github.com>