[FILECOIN][UXIT-3186] Refactor Navigation constants (#1728)
* refactor(navigation): Centralize and clarify navigation data

This commit refactors the navigation constants and types to improve clarity and maintainability.

* fix(ui): Update navigation menu grid layout

The previous `grid-flow-col` implementation forced navigation items to fill the grid vertically. This makes it vertically, the default grid behavior.

* style: Increase border radius on navigation menu panel

This commit updates the `roundedStyle` constant in the `NavigationMenuPanel` component from `rounded-xl` to `rounded-2xl`.

* refactor(nav): Elevate Community Hub to top-level navigation

Move the "Community Hub" link from the "Community" dropdown menu to the main header navigation bar. This change increases its visibility and makes it a primary navigation item.

* CR - Move variantMapping into constants

* CR - Use satisfies for variantMapping type annotation

* CR - Move navigation constants closer to component

* CR - Use import aliases

* CR - Prettify

* CR - Organize NavigationMenu components under own folder

* CR - Move hooks under own folder

* CR - Use local imports

* CR - Use labels

* CR - Add GitHub label

* CR - Add securityBugBounty label

* CR - Use labels for FILECOIN_FOUNDATION urls

* CR

* CR - Use FILECOIN_FOUNDATION_URL constant

---------

Co-authored-by: Mirha Masala <mirha@hey.com>